インデックスバッファの要素はu16なので修正

+++ include/toolbox/td/vulkan/td.h
@@ -68,9 +68,7 @@ namespace TB {
 						} texCoord;
 					};
 
-					Object(
-						const std::vector<Vertex>&,
-						const std::vector<unsigned>&);
+					Object(const std::vector<Vertex>&, const std::vector<u16>&);
 					~Object();
 
 					void Draw(VkCommandBuffer&);
